# api/swagger/swagger.yaml
swagger: '2.0'
info:
  version: 0.0.1
  title: FlockIt
host: 'localhost:3000'
basePath: /
tags:
  - {name: Service, description: 'Operations regarding the creation and administration of services.'}
  - {name: User, description: 'Operations regarding the creation and administration of users.'}
  - {name: Event, description: 'Operations regarding the creation and administration of events.'}
schemes:
  - http
  - https
consumes:
  - application/json
produces:
  - application/json
definitions:
  Service: {type: object, required: [fk_provider, tx_name], properties: {tx_cif: {type: string, example: A05732B}, fk_provider: {type: integer, format: uint32, example: 1}, tx_name: {type: string, example: 'Test service'}, fk_headquarters: {type: integer, format: uint32, example: 1}, url_main_image: {type: string}, url_main_video: {type: string}}}
  UsersArray: {type: array, items: {type: object, required: [tx_email, fk_place], properties: {tx_email: {type: string, example: rubenlopezlozoya12@gmail.com}, tx_name: {type: string, example: Ruben}, url_profile_image: {type: string, example: /good_url.img}, tx_bio: {type: string, example: 'Test bio'}, tx_gender: {type: string, example: M}, dt_birthday: {type: string, format: dateTime, example: '1993-04-08T22:00:00.000Z'}, fk_place: {type: integer, format: uint32, example: 1}, n_rating: {type: number, format: double, example: 3.2}}}}
  User: {type: object, required: [tx_email, fk_place], properties: {tx_email: {type: string, example: rubenlopezlozoya12@gmail.com}, tx_name: {type: string, example: Ruben}, url_profile_image: {type: string, example: /good_url.img}, tx_bio: {type: string, example: 'Test bio'}, tx_gender: {type: string, example: M}, dt_birthday: {type: string, format: dateTime, example: '1993-04-08T22:00:00.000Z'}, fk_place: {type: integer, format: uint32, example: 1}, n_rating: {type: number, format: double, example: 3.2}}}
  CreateUserResponse: {type: object, properties: {userId: {type: integer, format: uint32}}}
  CreateServiceResponse: {type: object, properties: {serviceId: {type: integer, format: uint32}}}
  Error: {required: [message], properties: {message: {type: string}}}
paths:
  /service: {x-swagger-router-controller: service, post: {tags: [service], description: 'Creates a service', operationId: create, parameters: [{in: body, name: body, description: 'Service object that needs to be added to the database', required: true, schema: {type: object, required: [fk_provider, tx_name], properties: {tx_cif: {type: string, example: A05732B}, fk_provider: {type: integer, format: uint32, example: 1}, tx_name: {type: string, example: 'Test service'}, fk_headquarters: {type: integer, format: uint32, example: 1}, url_main_image: {type: string}, url_main_video: {type: string}}}}], responses: {'200': {description: Success, schema: {type: object, properties: {serviceId: {type: integer, format: uint32}}}}, default: {description: Error, schema: {required: [message], properties: {message: {type: string}}}}}}}
  '/user/{userId}': {x-swagger-router-controller: user, get: {tags: [user], parameters: [{name: userId, in: path, required: true, type: integer, format: uint32}], description: 'Retrieves a user', operationId: findById, responses: {'200': {description: Success, schema: {type: array, items: {type: object, required: [tx_email, fk_place], properties: {tx_email: {type: string, example: rubenlopezlozoya12@gmail.com}, tx_name: {type: string, example: Ruben}, url_profile_image: {type: string, example: /good_url.img}, tx_bio: {type: string, example: 'Test bio'}, tx_gender: {type: string, example: M}, dt_birthday: {type: string, format: dateTime, example: '1993-04-08T22:00:00.000Z'}, fk_place: {type: integer, format: uint32, example: 1}, n_rating: {type: number, format: double, example: 3.2}}}}}, default: {description: Error, schema: {required: [message], properties: {message: {type: string}}}}}}}
  /swagger: {x-swagger-pipe: swagger_raw}